feat: добавление авторизации, централизованная обработка ошибок, валидация

+++ controllers/cards.js
@@ -1,8 +1,11 @@
 const errorConstants = require('http2').constants;
 const mongoose = require('mongoose');
 const Card = require('../models/card');
+const BadRequestError = require('../errors/BadRequestError');
+const NotFoundError = require('../errors/NotFoundError');
+const ForbiddenError = require('../errors/ForbiddenError');
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => {
@@ -12,71 +15,84 @@ module.exports.createCard = (req, res) => {
         .then((data) => res.status(errorConstants.HTTP_STATUS_CREATED).send(data))
         .catch((error) => {
           if (error instanceof mongoose.Error.DocumentNotFoundError) {
-            res.status(errorConstants.HTTP_STATUS_NOT_FOUND).send({ message: 'Карточка с указанным id не найдена' });
+            next(new NotFoundError('Карточка с указанным id не найдена'));
           } else {
-            res.status(errorConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+            next(error);
           }
         });
     })
     .catch((error) => {
       if (error instanceof mongoose.Error.ValidationError) {
-        res.status(errorConstants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Переданы некорректные данные при создании карточки' });
+        next(new BadRequestError('Переданы некорректные данные при создании карточки'));
       } else {
-        res.status(errorConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+        next(error);
       }
     });
 };
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .populate(['owner', 'likes'])
     .then((cards) => res.send(cards))
-    .catch(() => res.status(errorConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' }));
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail()
-    .then(() => res.send({ message: 'Карточка успешно удалена' }))
-    .catch((error) => {
-      if (error instanceof mongoose.Error.CastError) {
-        res.status(errorConstants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Некорректный id' });
-      } else if (error instanceof mongoose.Error.DocumentNotFoundError) {
-        res.status(errorConstants.HTTP_STATUS_NOT_FOUND).send({ message: 'Карточка с указанным id не найдена' });
-      } else {
-        res.status(errorConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .then((card) => {
+      if (!card.owner.equals(req.user._id)) {
+        throw new ForbiddenError('Вы не можете удалять чужие карточки');
       }
+      Card.findByIdAndRemove(req.params.cardId)
+        .orFail()
+        .then(() => res.send({ message: 'Карточка успешно удалена' }))
+        .catch((error) => {
+          if (error instanceof mongoose.Error.CastError) {
+            next(new BadRequestError('Некорректный id'));
+          } else if (error instanceof mongoose.Error.DocumentNotFoundError) {
+            next(new NotFoundError('Карточка с указанным id не найдена'));
+          } else {
+            next(error);
+          }
+        })
+        .catch((error) => {
+          if (error instanceof mongoose.Error.DocumentNotFoundError) {
+            next(new NotFoundError('Карточка с указанным id не найдена'));
+          } else {
+            next(error);
+          }
+        });
     });
 };
 
-module.exports.addLike = (req, res) => {
+module.exports.addLike = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
     .orFail()
     .populate(['owner', 'likes'])
     .then((card) => res.send(card))
     .catch((error) => {
       if (error instanceof mongoose.Error.CastError) {
-        res.status(errorConstants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Некорректный id' });
+        next(new BadRequestError('Некорректный id'));
       } else if (error instanceof mongoose.Error.DocumentNotFoundError) {
-        res.status(errorConstants.HTTP_STATUS_NOT_FOUND).send({ message: 'Карточка с указанным id не найдена' });
+        next(new NotFoundError('Карточка с указанным id не найдена'));
       } else {
-        res.status(errorConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+        next(error);
       }
     });
 };
 
-module.exports.deleteLike = (req, res) => {
+module.exports.deleteLike = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
     .orFail()
     .populate(['owner', 'likes'])
     .then((card) => res.send(card))
     .catch((error) => {
       if (error instanceof mongoose.Error.CastError) {
-        res.status(errorConstants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Некорректный id' });
+        next(new BadRequestError('Некорректный id'));
       } else if (error instanceof mongoose.Error.DocumentNotFoundError) {
-        res.status(errorConstants.HTTP_STATUS_NOT_FOUND).send({ message: 'Карточка с указанным id не найдена' });
+        next(new NotFoundError('Карточка с указанным id не найдена'));
       } else {
-        res.status(errorConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+        next(error);
       }
     });
 };
